requirements.txt
Flask>=1.1.1
python-dotenv>=1.0.0
python-dotenv==1.2.1